Imporve docs
