+++ lib/src/models/models.dart
@@ -2,9 +2,9 @@ export 'enum.dart';
 export 'account/account.dart';
 export 'action/action.dart';
 export 'action/action_result.dart';
+export 'address/address.dart';
 export 'auth/auth.dto.dart';
 export 'collectible/collectible.dart';
-export 'collection/address.dart';
 export 'collection/collection.dart';
 export 'relay/relay.dart';
 export 'smart_wallet/smart_wallet.dart';
